refactor(direct-control): extract production choice types
Refs: openspec/changes/civ7-support-direct-control-modularization

Moves production-choice input, command payload, and result public type ownership into src/play/operations/production-choice.ts while keeping public facade type re-exports in index.ts.

Leaves diplomacy/narrative closeout, public procedure schemas, operation/proof telemetry, hotseat runtime proof, AI ingestion, CLI semantic projection, Effect/oRPC procedure-core work, and Task 2.9.4 matrix-row acceptance pending.

Validation:
- bun run --cwd packages/civ7-direct-control test -- production-choice.test.ts public-api.test.ts
- bun run --cwd packages/cli test -- game.play.production.test.ts
- bun run --cwd packages/civ7-direct-control check
- bun run --cwd packages/civ7-direct-control test
- bun run --cwd packages/civ7-direct-control build
- bun run openspec -- validate civ7-support-direct-control-modularization --strict
- bun run resources:status
- bun run check:cli
- bun run test:cli:play
- bun run --cwd apps/mapgen-studio check
- git diff --check
- ownership/no-cycle scans for production-choice type definitions and facade executable imports

Local package/type relocation proof only; no runtime, live-game, AI-on-hotseat, telemetry, semantic CLI, or procedure-core proof claimed.
