+++ openspec/changes/civ7-support-direct-control-modularization/tasks.md
@@ -493,6 +493,13 @@ runtime/direct-control claims.
         procedure schemas, operation/proof telemetry, hotseat runtime proof, AI
         ingestion, CLI semantic projection, Effect/oRPC procedure-core work,
         and Task 2.9.4 matrix-row acceptance pending.
+  - [x] 4.8.27 Extract production-choice type owner while keeping public
+        facade type re-exports in `index.ts`, preserving production-choice
+        input, command payload, and result public contracts while leaving
+        diplomacy/narrative closeout, public procedure schemas,
+        operation/proof telemetry, hotseat runtime proof, AI ingestion, CLI
+        semantic projection, Effect/oRPC procedure-core work, and Task 2.9.4
+        matrix-row acceptance pending.
 - [ ] 4.9 Import or explicitly cite oRPC architecture authority before any
       oRPC implementation; current support branch has no tracked
       `.agents/skills/civ7-orpc-control-architecture` or
