+++ openspec/changes/civ7-support-direct-control-modularization/workstream/workstream-record.md
@@ -1195,6 +1195,18 @@ All future agent waves must be framed before delegation:
   semantic projection, Effect/oRPC procedure-core work, and Task 2.9.4
   matrix-row acceptance remain pending. This is local package/type relocation
   proof only, not runtime/live-game proof.
+- Direct-control production-choice types slice: completed as a narrow public
+  type ownership relocation. It moves production-choice input, command payload,
+  and result public type ownership into
+  `src/play/operations/production-choice.ts` while keeping public facade type
+  re-exports in `index.ts`. This preserves public type contracts and does not
+  change production-choice source strings, wrapper orchestration, validation,
+  approval-first behavior, production postconditions, or request semantics.
+  Diplomacy/narrative closeout, public procedure schemas, operation/proof
+  telemetry, hotseat runtime proof, AI ingestion, CLI semantic projection,
+  Effect/oRPC procedure-core work, and Task 2.9.4 matrix-row acceptance remain
+  pending. This is local package/type relocation proof only, not
+  runtime/live-game proof.
 - Direct-control diplomacy closeout source slice: completed as a narrow
   operation source relocation. It moves only the App UI diplomacy response
   closeout command builder and embedded closeout source into
